+++ CHANGELOG.md
@@ -8,8 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Tests for magic config variables: `window.__HC_LAUNCHER_ENV__` and `window.__HC_ZOME_CALL_SIGNER__`
 - Added `SignalType` enum
 ### Fixed
+- Fix types `AppInfoStatus`, `PausedAppReason` and `DisabledAppReason`
 ### Changed
+- Rename `InstalledAppInfoStatus` to `AppInfoStatus`
 ### Removed
+- Remove unused type `DeactivationReason`
 
 ## 2025-04-07: v0.19.0-rc.0
 ### Added